fix: handle objects and prefixes more consistently
This improves several situations where we had either objects or prefixes
and were making decisions thinking that we only had the other -- for
example thinking that prefixes mattered but objects are what we're actually
deciding on.

That comes down to filtering objects earlier so that we can know if the 
objects we have are meaningful, and not bailing if we've run out of
prefixes when objects are still helpful.